vpn_server_prod/firehol/app/requirements.txt
requests
dnspython
backoff
